requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
pytz>=2023.3
